Added red error messages to the bulleted "featured" fields that are normally listed on procedures on the part pages (Dashboard and Event Stream). This way, you don't have to drill in to see errors.  Also made the disposition in the dashboard show with a red background in the case where the disposition it self is not correct (set to Pass, but there are errors in the fields).
Fixed a problem with PDFs getting stuck on certain UTF-8 characters. The entire Eventstream would be missing in the PDF past the offending character (a U+2019 appostrophie in this case). There were a series of things going wrong. Ultimately I needed to remove the function translateUtf8ToIso() which was converting to ISO 8859-1 from UTF-8 before feeding to the PDF. This is not needed now since the PDF can handle UTF-8.  Also I stopped using the htmLawed(clean_ms_char) feature since it was corrupting the UTF-8 code. I finally added a final command to strip some known UTF-8 characters that were killing the PDF generation.

Plus a little code cleanup.