Added the QR Code Upload feature where you can scan a code on the comment edit page and this will open a pasword-less connection on the phone browser to let you take and select photos for uploading. The password-less access is a /utils/qrupload. There are now two new tables in the database as a result. There is also another task in the MaintenanceTaskRunner process to cleanup old records in the tables that are left behind by the upload process.
Changed the document outputToBrowser handler so that now if asked to return a thumbnail for a document with no thumbnail, it will return an icon instead. This we needed to fix the problem of blank images for non-photo documents on the commenteditview page.

Unrelated to the QR code thing: Fixed Notice in the ImportProcess. This was causing a fail with PartForgeLink in my testing environment.