First full commit of changes that introduce the new field type "attachment" which lets you attach files and photos into the versioned form.
Other changes: Fixed several bugs in the popup to show unversioned changes. This included a bug always showing a change to the Part Number. Changed the "You must save this record..." handling when creating a comment or a new component so that it doesn't ask you to fix errors. Improved efficiency of calculating resursive validation errors. Fixed bug where simply saving before creating a subcomponent or adding a field attachment considered that to be two edits and so showed the "Edited Once" link for unversioned changes. Slight tweak to boolean print field logic so "" is returned as empty instead of No.